requirements.txt
# ❌ REMOVE hardcoded values
email="admin@gmail.com"
password_hash=hash_password("admin123")

# ✅ REPLACE WITH
admin_email    = os.getenv("ADMIN_EMAIL",    "admin@gmail.com")
admin_password = os.getenv("ADMIN_PASSWORD", "admin123")

# full fixed function:
@app.on_event("startup")
def create_default_admin():
    db = SessionLocal()
    admin_email    = os.getenv("ADMIN_EMAIL",    "admin@gmail.com")
    admin_password = os.getenv("ADMIN_PASSWORD", "admin123")
    try:
        if not db.query(User).filter(User.email == admin_email).first():
            db.add(User(
                email=admin_email,
                password_hash=hash_password(admin_password)
            ))
            db.commit()
            print("✅ Admin created.")
    finally:
        db.close()
```

---

### File 3 — `requirements.txt` — clean it up

Delete everything from the `pip install annotated-doc...` line downward. Keep only the clean list at the top. Also add `gunicorn` and `psycopg2-binary`:
```
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.133.1
filelock==3.24.3
greenlet==3.3.2
gunicorn==21.2.0
h11==0.16.0
idna==3.11
joblib==1.5.3
numpy==2.2.6
packaging==26.0
pandas==2.3.3
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic==2.12.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.5.0
python-multipart==0.0.22
pytz==2025.2
scikit-learn==1.7.2
scipy==1.15.3
shap==0.49.1
SQLAlchemy==2.0.47
starlette==0.52.1
torch==2.10.0
tqdm==4.67.3
typing_extensions==4.15.0
uvicorn==0.41.0